BUG FIX: When changing the debug tab, debug information was also updated to that editor even though it was not used to compile the scene. Smaller code improvements.
