+++ App/demos/shadow_mapping_pcss.tech
@@ -99,9 +99,9 @@ shader vert vs_render {
 	void main ()
 	{
 		// transform vertex to camera projection space
-		gl_Position = camera.viewProj * in_pos;
+		<<<gl_Position>>> = camera.viewProj * in_pos;
 		// transform vertex to light projection space
-		out_lpos = light.viewProj * in_pos;
+		<<<out_lpos>>> = light.viewProj * in_pos;
 		// pass vertex color
 		out_col = in_col;
 	}
@@ -224,7 +224,7 @@ shader frag fs_render {
 	
 	void main () {
 		// pass color
-		color = in_col;
+		<<<color>>> = in_col;
 		
 		//vec3 lpos = (in_lpos.xyz / in_lpos.w) * 0.5 + 0.5;
 		//color.rgb = normalize(cross(dFdx(lpos),dFdy(lpos)));
@@ -248,7 +248,7 @@ shader frag fs_render {
 		
 		//const vec2 r = 10.0 / textureSize(shadowmap, 0);
 		//float shadow = PCF((r.x + r.y) * 0.5, lpos, zChange);
-		float shadow = PCSS(surfBias, lpos);
+		float <<<shadow>>> = PCSS(surfBias, lpos);
 			
 		if (shadow == 0.0)
 			return;
