+++ App/src/GraphicControl.cs
@@ -1,5 +1,4 @@
 ﻿using App;
-using OpenTK.Graphics.OpenGL4;
 using System;
 using System.Collections.Generic;
 using System.Linq;
